+++ pom.xml
@@ -6,7 +6,7 @@
                              http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>az.cybernet.internship</groupId>
-  <artifactId>dictionary-service</artifactId>
+  <artifactId>dictionaryEntity-service</artifactId>
   <version>1.0.0-SNAPSHOT</version>
 
   <parent>
