+++ src/main/java/az/cybernet/internship/dictionary/DictionaryServiceApplication.java
@@ -3,7 +3,7 @@ package az.cybernet.internship.dictionary;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-@MapperScan()
+@MapperScan("az.cybernet.internship.dictionary.mapper")
 @SpringBootApplication
 public class DictionaryServiceApplication {
     public static void main(String[] args) {
