+++ src/main/java/az/cybernet/internship/dictionary/dto/DictionaryResponse.java
@@ -1,11 +1,14 @@
-package DictionaryDto;
+package az.cybernet.internship.dictionary.dto;
 
 import lombok.Data;
 
+import java.util.UUID;
+
 @Data
 public class DictionaryResponse {
-    private String id;
-    private String category;
+
+
+    private UUID id;
     private String value;
     private String description;
     private Boolean isActive;
