--- src/main/java/model/Dictionary.java
@@ -1,16 +1,0 @@
-package model;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class Dictionary {
-    private String id;
-    private String category;
-    private String value;
-    private String description;
-    private Boolean isActive;
-    private LocalDateTime deletedAt;
-
-}
