+++ src/main/resources/DictionaryMyBatis.xml
@@ -3,10 +3,10 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="az.cybernet.internship.dictionary.mapper.IDictionaryMapper">
+<mapper namespace="az.cybernet.internship.dictionaryEntity.az.cybernet.internship.dictionary.mapper.IDictionaryMapper">
 
-    <select id="findAll" resultType="az.cybernet.internship.dictionary.model.Dictionary">
-        SELECT * FROM dictionary
+    <select id="getDictionaries" resultType="az.cybernet.internship.dictionaryEntity.az.cybernet.internship.dictionary.model.DictionaryEntity">
+        SELECT * FROM dictionaryEntity
         WHERE 1=1
         <if test="id != null">AND id = #{id}</if>
         <if test="value != null">AND LOWER(value) LIKE CONCAT('%', LOWER(#{value}), '%')</if>
@@ -14,29 +14,29 @@
         <if test="limit != null">LIMIT #{limit}</if>
     </select>
 
-    <select id="findById" resultType="az.cybernet.internship.dictionary.model.Dictionary">
-        SELECT * FROM dictionary WHERE id = #{id}
+    <select id="findById" resultType="az.cybernet.internship.dictionaryEntity.az.cybernet.internship.dictionary.model.DictionaryEntity">
+        SELECT * FROM dictionaryEntity WHERE id = #{id}
     </select>
 
     <insert id="insert">
-        INSERT INTO dictionary (id, category, value, description, is_active)
+        INSERT INTO dictionaryEntity (id, category, value, description, is_active)
         VALUES (#{id}, #{category}, #{value}, #{description}, #{isActive})
     </insert>
 
     <update id="update">
-        UPDATE dictionary
+        UPDATE dictionaryEntity
         SET category = #{category}, value = #{value}, description = #{description}, is_active = #{isActive}
         WHERE id = #{id}
     </update>
 
     <update id="softDelete">
-        UPDATE dictionary
+        UPDATE dictionaryEntity
         SET is_active = false, deleted_at = #{timestamp}
         WHERE id = #{id}
     </update>
 
     <update id="restore">
-        UPDATE dictionary
+        UPDATE dictionaryEntity
         SET is_active = true, deleted_at = null
         WHERE id = #{id}
     </update>
