+++ src/main/resources/application.yml
@@ -21,5 +21,5 @@ springdoc:
 
     mybatis:
       mapper-locations: classpath:mappers/*.xml
-      type-aliases-package: com.example.model
+      type-aliases-package: com.example.az.cybernet.internship.dictionary.model
